+++ cmd/hidden_lake/applications/messenger/pkg/app/init_service.go
@@ -10,15 +10,14 @@ import (
 	hlm_settings "github.com/number571/go-peer/cmd/hidden_lake/applications/messenger/pkg/settings"
 	"github.com/number571/go-peer/cmd/hidden_lake/applications/messenger/web"
 	"github.com/number571/go-peer/pkg/logger"
-	"github.com/number571/go-peer/pkg/queue_set"
 	"golang.org/x/net/websocket"
 )
 
-func (p *sApp) initIncomingServiceHTTP(pQueuePusher queue_set.IQueuePusher) {
+func (p *sApp) initIncomingServiceHTTP() {
 	mux := http.NewServeMux()
 	mux.HandleFunc(
 		hlm_settings.CPushPath,
-		handler.HandleIncomigHTTP(p.fHTTPLogger, p.fConfig, p.fDatabase, pQueuePusher),
+		handler.HandleIncomigHTTP(p.fHTTPLogger, p.fConfig, p.fDatabase),
 	) // POST
 
 	p.fIncServiceHTTP = &http.Server{
@@ -28,7 +27,7 @@ func (p *sApp) initIncomingServiceHTTP(pQueuePusher queue_set.IQueuePusher) {
 	}
 }
 
-func (p *sApp) initInterfaceServiceHTTP(pQueuePusher queue_set.IQueuePusher) {
+func (p *sApp) initInterfaceServiceHTTP() {
 	mux := http.NewServeMux()
 	mux.Handle(hlm_settings.CStaticPath, http.StripPrefix(
 		hlm_settings.CStaticPath,
@@ -37,13 +36,13 @@ func (p *sApp) initInterfaceServiceHTTP(pQueuePusher queue_set.IQueuePusher) {
 
 	cfgWrapper := config.NewWrapper(p.fConfig)
 
-	mux.HandleFunc(hlm_settings.CHandleIndexPath, handler.IndexPage(p.fHTTPLogger, p.fConfig))                                        // GET, POST
-	mux.HandleFunc(hlm_settings.CHandleFaviconPath, handler.FaviconPage(p.fHTTPLogger, p.fConfig))                                    // GET
-	mux.HandleFunc(hlm_settings.CHandleAboutPath, handler.AboutPage(p.fHTTPLogger, p.fConfig))                                        // GET
-	mux.HandleFunc(hlm_settings.CHandleSettingsPath, handler.SettingsPage(p.fHTTPLogger, cfgWrapper))                                 // GET, PATCH, PUT, POST, DELETE
-	mux.HandleFunc(hlm_settings.CHandleFriendsPath, handler.FriendsPage(p.fHTTPLogger, cfgWrapper))                                   // GET, POST, DELETE
-	mux.HandleFunc(hlm_settings.CHandleFriendsChatPath, handler.FriendsChatPage(p.fHTTPLogger, p.fConfig, p.fDatabase, pQueuePusher)) // GET, POST, PUT
-	mux.HandleFunc(hlm_settings.CHandleFriendsUploadPath, handler.FriendsUploadPage(p.fHTTPLogger, p.fConfig))                        // GET
+	mux.HandleFunc(hlm_settings.CHandleIndexPath, handler.IndexPage(p.fHTTPLogger, p.fConfig))                          // GET, POST
+	mux.HandleFunc(hlm_settings.CHandleFaviconPath, handler.FaviconPage(p.fHTTPLogger, p.fConfig))                      // GET
+	mux.HandleFunc(hlm_settings.CHandleAboutPath, handler.AboutPage(p.fHTTPLogger, p.fConfig))                          // GET
+	mux.HandleFunc(hlm_settings.CHandleSettingsPath, handler.SettingsPage(p.fHTTPLogger, cfgWrapper))                   // GET, PATCH, PUT, POST, DELETE
+	mux.HandleFunc(hlm_settings.CHandleFriendsPath, handler.FriendsPage(p.fHTTPLogger, cfgWrapper))                     // GET, POST, DELETE
+	mux.HandleFunc(hlm_settings.CHandleFriendsChatPath, handler.FriendsChatPage(p.fHTTPLogger, p.fConfig, p.fDatabase)) // GET, POST, PUT
+	mux.HandleFunc(hlm_settings.CHandleFriendsUploadPath, handler.FriendsUploadPage(p.fHTTPLogger, p.fConfig))          // GET
 
 	mux.Handle(hlm_settings.CHandleFriendsChatWSPath, websocket.Handler(handler.FriendsChatWS))